move load config to test utils
